refactor(upgrade): split runners.rs god-file into modules
